+++ legacy/src/widgets/component-new-edit/components/response-format/pairing/response-format-pairing.jsx
@@ -10,6 +10,7 @@ import {
 import GenericOption from '../../../../../forms/controls/generic-option';
 import Select from '../../../../../forms/controls/select';
 import Dictionary from '../../../../../utils/dictionary/dictionary';
+import { requiredSelect } from '../../../../../utils/validation/validate-rules';
 import { getCurrentSelectorPath } from '../../../../../utils/widget-utils';
 import { CodesLists } from '../../../../codes-lists';
 
@@ -105,6 +106,7 @@ function ResponseFormatPairing({
         component={Select}
         label={Dictionary.pairingSourceVariable}
         required
+        validate={[requiredSelect]}
       >
         <GenericOption key="" value="">
           {Dictionary.selectBasedOn}