+++ next/src/components/personalization/form/PersonalizationForm.tsx
@@ -14,7 +14,9 @@ import {
 import { getFileFromParseResult } from '@/api/utils/personalization'
 import Button, { ButtonStyle } from '@/components/ui/Button'
 import ButtonIcon, { ButtonIconStyle } from '@/components/ui/ButtonIcon'
+import CsvViewerTable from '@/components/ui/CsvViewerTable'
 import DialogButton from '@/components/ui/DialogButton'
+import UploadMessageTile from '@/components/ui/UploadMessageTile'
 import Field from '@/components/ui/form/Field'
 import FormInput from '@/components/ui/form/FormInput'
 import RadioGroup from '@/components/ui/form/RadioGroup'
@@ -28,8 +30,6 @@ import {
   UploadMessage,
 } from '@/models/personalizationQuestionnaire'
 
-import UploadMessageTile from '../overview/UploadMessageTile'
-import CsvViewerTable from './CsvViewerTable'
 import JsonViewer from './JsonViewer'
 
 interface PersonalizationFormProps {