+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [3.2.5](https://github.com/InseeFr/Pogues/releases/tag/3.2.4) - 2026-06-09
+
+### Added
+
+- Import codes list from a CSV file when creating or editing a codes list.
 
 #### Fixed